backend/requirements.txt
# FastAPI and Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
motor==3.3.2
pymongo==4.6.1
beanie==1.24.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# Environment & Configuration
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# ML & AI
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4

# HTTP Client
httpx==0.26.0
requests==2.31.0

# Utilities
python-dateutil==2.8.2
email-validator==2.1.0
PyMuPDF==1.23.21
python-docx==1.1.0

# Note: llama-cpp-python removed for easier installation
# You can add it later if needed for LLM features
